pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "website"
version = "0.1.0"
[tool.poetry.dependencies]
flask = "^1.1"
python = "^3.7"
pyfiglet = "^0.8.0"
IMDbPY = "^6.8"
[tool.poetry.dev-dependencies]
